Use fastify plugin metadata or an explicit passed in name option for a plugin's name
`getName` in avvio gets a string name for each plugin registered to use when reporting debug info. Doing some profiling of a big fastify application, I noticed that a lot of time was spent getting the name of avvio plugins in the getName function. For me, require.cache was very big, and the number of registered plugins was very big, so the O(n^2) search for each file in the require cache was actually taking a good long while. Also, getName relying on the require cache is undesirable as ESM becomes more and more common -- we should try to move to an approach that doesn't need that.

Fortunately, a lot of plugins registered with avvio already report their names! We were somewhat accidentally using `getName` to get the name of a lot of plugins that already have fastify metadata set, or pass an explicit name in with their options. We can avoid the nasty performance pitfall by just accepting the incoming names inside avvio, and then falling back on one of the much faster last-resort strategies if plugin names aren't set.

It is kind of unfortunate that we can't use the plugin's file path as a nice, human friendly name for a plugin that isn't otherwise named, but I think the performance issue merits its removal, and I will add documentation to fastify to indicate that naming your own plugins is a good idea.